Fix project defaulting for cascading directories
Before, the default project was global and could not be overridden, and
specifying projects on the command line was fragile.

Fix this to the project can be defaulted at any level, and starting a
workspace from any subdir of a project defaults to that project.  In
addition, when completely outside a tree, the global default works.
Finally, when within another tree, specifying the project on the command
line works.

Add verbose options to see what's being loaded when.