die - Make work in subshells
The die() function is very useful, but it could not be used well in
functions.  If the function was run from a script, the entire script
would abort, not just the function.  And if the function was run from
the commandline, then it would keep running, and not abort at all.

To work around this, Make die() detect when it's in a subshell, and
actually exit.  This allows a function run from the commandline to run
things in a subshell, allowing die() to work properly.  For example:

dietest() {
	(
		command1 || die "command1 failed"
		command2 || die "command2 failed"
		command3 || die "command3 failed"
	)
	return $?
}

When run by the user, this will abort if any command fails, but not exit
the user's shell, and properly set the exit code.  If run in a script,
this will abort the function, and return failure.  If it is desired that
the entire script abort, the return can be replaced with this:

	local val=$?
	[[ "${val}" == "0" ]] || die
	return ${val}