gtws-docker - Start a workspace in docker
The gtws-docker command can start a workspace in Docker.  The docker
image must be alreay set up for this project, but the container will be
unique to this workspace.

In addition, startws is given a --init option that will initialize the
path to the workspace for the given project from the default settings.
This is because docker mounts the workspace into a (potentially) blank
tree, so config doesn't exist.