gtws-mirror - setup a local mirror for a project
This will create a local mirror of a project, incluing all it's
submodules.  It can be run from inside or outside a workspace, but it
needs a checkout from upstream to operate on if it's going to mirror the
main repo.

In submodule only mode, it will work with a local main repo mirror, and
add mirrors for any submodules that are not currently mirrored.