gtws-clone - allow to work with remote repos
gtws_opv is now gtws_opvn, and takes the WS name as well.  If it is a
remote repo, then the returned value has the name on the end to index
into the remote WS.  Otherwise, it works like before.

gtws-clone can now work with remote repos and local submodule mirrors.