mkws - Add per-project and per-repo extras
mkws / gtws-clone now look for a directory tree in ${opv%/git}/extra to
copy into the workspace and/or repo(s).  The structure looks like this:

extra/ws		- Stuff to copy into the top level of the
			  workspace
extra/repo/<reponame>	- Stuff to copi into the top level of the repo

Everything is copied exactly, including dotfiles, dotdirs, files and
dirs.

Note that, for repos, the copy comes from the opv of the repo, not from
the opv of the workspace.